+++ src/components/layout/header/AvatarDropdownMenu.module.css
