+++ src/features/workspace/components/ArticleContent/editor-styles.scss
