+++ src/features/workspace/components/Thumbnail.tsx
