+++ src/features/workspace/components/editor/customExtensions/CustomPasteImageExtension.tsx
