+++ src/lib/api/article/publish.ts
@@ -1,4 +1,4 @@
-import { privateFetch } from "@/lib/fetch/privateFetch"
+import { privateFetch } from "@/lib/api/fetch/privateFetch"
 
 interface PublishRequest{
     title: string
