+++ src/lib/api/auth/logout.ts
@@ -1,7 +1,7 @@
 import { store } from "@/app/store/store"
 import { clear } from "@/features/auth/store/authSlice"
 import { clearUser } from "@/features/user/store/userSlice"
-import { privateFetch } from "@/lib/fetch/privateFetch"
+import { privateFetch } from "@/lib/api/fetch/privateFetch"
 
 interface LogoutResponse {
     message: string
