+++ src/lib/api/folder/renameFolder.ts
@@ -1,4 +1,4 @@
-import { privateFetch } from "@/lib/fetch/privateFetch"
+import { privateFetch } from "@/lib/api/fetch/privateFetch"
 
 interface RenameFolderRequest{
     folderId: string
